// app/types/supabase.ts
export type Database = {
  public: {
    Tables: {
      Game: {
        Row: {
          created_at: string
          endPrice: number | null
          fid: string
          gameEndTimeStamp: string
          gameStartTimeStamp: string
          id: string
          position: string
          startPrice: number
          status: string
          token: string
          userId: string
          wager: number
        }
        Insert: {
          created_at?: string
          endPrice?: number | null
          fid: string
          gameEndTimeStamp?: string
          gameStartTimeStamp?: string
          id?: string
          position: string
          startPrice: number
          status?: string
          token: string
          userId?: string
          wager: number
        }
        Update: {
          created_at?: string
          endPrice?: number | null
          fid?: string
          gameEndTimeStamp?: string
          gameStartTimeStamp?: string
          id?: string
          position?: string
          startPrice?: number
          status?: string
          token?: string
          userId?: string
          wager?: number
        }
        Relationships: [
          {
            foreignKeyName: "public_Game_fid_fkey"
            columns: ["fid"]
            isOneToOne: false
            referencedRelation: "User"
            referencedColumns: ["fid"]
          },
        ]
      }
      User: {
        Row: {
          address: string | null
          balance: number
          created_at: string
          fid: string
          id: string
        }
        Insert: {
          address?: string | null
          balance: number
          created_at?: string
          fid: string
          id?: string
        }
        Update: {
          address?: string | null
          balance?: number
          created_at?: string
          fid?: string
          id?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never